+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Firefox.
 
 - General: don't fail if search result container isn't detected, just write to
   console.
+- YouTube: fix navigation in non-grid pages
 
 ## 0.4.22